feat: add security config

+++ src/main/java/com/github/sparsick/workshop/Book.java
@@ -53,11 +53,11 @@ public class Book implements Serializable{
     }
 
         
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
